+++ src/rpc_addr.ml
@@ -16,20 +16,20 @@ let of_string s = s
 let to_string t = t
 
 let host t =
-  match String.split_on_char ':' t with
-  | [h; _] ->
+  match Host_port.split t with
+  | Error _ -> None
+  | Ok (h, _) ->
       let h = String.trim h in
       if h <> "" then Some h else None
-  | _ -> None
 
 let port t =
-  match String.split_on_char ':' t with
-  | [_; p] -> (
+  match Host_port.split t with
+  | Error _ -> None
+  | Ok (_, p) -> (
       try
         let p = int_of_string (String.trim p) in
         if p > 0 && p < 65536 then Some p else None
       with _ -> None)
-  | _ -> None
 
 let to_endpoint t =
   let trimmed = String.trim t in